requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.6                # via pip-tools
dj-database-url==0.4.2
django-cors-headers==2.0.2
django-extensions==1.7.5
django-jsonfield==1.0.1
django==1.10.4
djangorestframework==3.5.3
first==2.0.1              # via pip-tools
flake8==3.3.0
gunicorn==19.7.0
httplib2==0.9.2           # via oauth2client
jsonpickle==0.9.4
mccabe==0.6.1             # via flake8
oauth2client==4.0.0
olefile==0.44             # via pillow
pillow==4.0.0
pip-tools==1.8.0
psycopg2==2.6.2
py==1.4.31                # via pytest
pyasn1-modules==0.0.8     # via oauth2client
pyasn1==0.1.9             # via oauth2client, pyasn1-modules, rsa
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pytest-django==3.1.2
pytest==3.0.5             # via pytest-django
requests==2.12.3
rsa==3.4.2                # via oauth2client
six==1.10.0               # via django-extensions, oauth2client, pip-tools
whitenoise==3.2.3
